+++ src/global.d.ts
@@ -5,26 +5,24 @@ import type {Linter} from 'eslint';
 import type {PublicApi} from 'stylelint';
 import type {MwConfig, LintSource} from './codemirror';
 
-interface LuaNode {
-	name: string;
-	range: [number, number];
-}
-interface luaparse {
-	defaultOptions: {luaVersion: string};
-	parse(s: string): {globals: LuaNode[]};
-	SyntaxError: new () => {message: string, index: number};
-}
-
 declare global {
 	module '/*' {
 		export {CodeMirror6};
 		export type {MwConfig, LintSource};
 	}
 
+	interface LuaReport {
+		line: number;
+		column: number;
+		end_column: number;
+		msg: string;
+		severity: 1 | 2;
+	}
+
 	const eslint: {
 		Linter: new () => Linter;
 	};
 	const stylelint: PublicApi;
-	const luaparse: luaparse;
+	const luacheck: {queue(s: string): Promise<LuaReport[]>};
 	const Parser: Parser;
 }